+++ NEWS.md
@@ -3,6 +3,7 @@
 * Added possibility to maintain old itemnames with `spss_swap()` function
 * Added the possibility of multiple group comparisons to `flex_table1()` (recommended by Rémi and Taylor)
 * Added the function `pretty_cm()` for plotting of caret::confusionMatrix()
+* Added the function `get_ICD_10_cats()` that returns ICD-10 F Diagnoses categories (thanks to Dirk Edelbuettel)
 
 # datscience 0.2.2